refactor(types): Factor out user role types
Small step in setting up types for the smartcard auth interface.
