+++ libs/types/src/index.ts
@@ -10,6 +10,7 @@ export * from './image';
 export * from './numeric';
 export * from './precinct_selection';
 export * from './result';
+export * from './smartcard_auth';
 export * from './tallies';
 export * from './user_session';
 export * from './voting_method';
